fix: 날짜 포맷 오류 해결
